+++ packages/nuxt/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @posthog/nuxt
 
+## 1.6.0
+
+### Minor Changes
+
+- [#3302](https://github.com/PostHog/posthog-js/pull/3302) [`fc5589f`](https://github.com/PostHog/posthog-js/commit/fc5589fcc51bd53ba818822831867d3c00d83a11) Thanks [@dmarticus](https://github.com/dmarticus)! - preserve $set_once semantics in local flag evaluation cache
+  (2026-04-07)
+
+### Patch Changes
+
+- Updated dependencies [[`fc5589f`](https://github.com/PostHog/posthog-js/commit/fc5589fcc51bd53ba818822831867d3c00d83a11)]:
+  - posthog-js@1.365.0
+  - @posthog/core@1.25.0
+  - posthog-node@5.29.0
+
+## 1.5.103
+
+### Patch Changes
+
+- Updated dependencies [[`b25b689`](https://github.com/PostHog/posthog-js/commit/b25b68967f7e85317e2aacb8ecc4dc66a95095eb)]:
+  - posthog-js@1.364.7
+
+## 1.5.102
+
+### Patch Changes
+
+- Updated dependencies [[`a01a3d5`](https://github.com/PostHog/posthog-js/commit/a01a3d55dc134b1b269be58c7922ce3780c57fc5), [`68cd4e5`](https://github.com/PostHog/posthog-js/commit/68cd4e5f1133b95bcb87c382a0ae5f1bcb96903d)]:
+  - @posthog/core@1.24.6
+  - posthog-js@1.364.6
+  - posthog-node@5.28.11
+
 ## 1.5.101
 
 ### Patch Changes